Finished sorting files a bit, organized now. Compiles and runs and connects to app
