+++ OASMan_ESP32/src/bitmaps.h
@@ -1,3 +1,6 @@
+#ifndef bitmaps_h
+#define bitmaps_h
+
 //https://www.dcode.fr/binary-image
 
   //regex for website output to array:
@@ -95,3 +98,6 @@ static const unsigned char PROGMEM logo_bmp_airtekk[] =
 0b00000000, 0b00000000, 0b00000000, 0b00000000, 0b00000000, 0b00000000, 0b00000000, 0b00000000, 0b00000000, 0b00000000, 0b00000000, 0b00000000, 0b00000000, 0b00000000, 0b00000000, 0b00000000, 
 0b00000000, 0b00000000, 0b00000000, 0b00000000, 0b00000000, 0b00000000, 0b00000000, 0b00000000, 0b00000000, 0b00000000, 0b00000000, 0b00000000, 0b00000000, 0b00000000, 0b00000000, 0b00000000, 
 };
+
+
+#endif
